requirements.txt
jsonrpclib.pelix
pika
pymongo
requests
redis
lxml
dateutil
newspaper3k
sklean
tensorflow
pandas
watchdog
keras
